Convert to float

+++ src/hittable.cuh
@@ -5,7 +5,7 @@
 struct hit_record {
   point3 p;
   vec3 normal;
-  double t;
+  float t;
   bool front_face;
 
   __device__ inline void set_face_normal(const ray& r, const vec3& outward_normal) {
@@ -16,6 +16,6 @@ struct hit_record {
 
 class hittable {
   public:
-    __device__ virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
+    __device__ virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
 };
 #endif
